consolidate code in wsgi.py
